pyproject.toml
[project]
name = "withorbit"
version = "0.1.0"
description = "Generate mnemonic medium spaced repetition prompts with Claude and embed them using Orbit"
readme = "README.md"
authors = [
    { name = "Eliad Cohen", email = "eliadcohen@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.1",
    "markdown-it-py>=3.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
]

[project.scripts]
withorbit = "withorbit.cli:cli"

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = ["pytest>=8.0"]
